requirements.txt
functions-framework==3.*
google-generativeai==0.6.0
google-cloud-firestore==2.16.1
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1